Add files via upload
Adds the registration, password confirmation log-in/log-out pages and a simple session.  Includes a function to theoretically email the password registration link to employees but it might require setting up a local SMTP mail server to work.  User input still needs some validation on the client side but the server side validation is done.  Tweaked a few things in other files so that the sessions, hashing and error messages would work.  Still need to do something for the "Forgot Password?" option at login.